requirements.txt
# RareDx — Pediatric Rare Disease Intelligence
# Pinned requirements for reproducibility
# Install: pip install -r requirements.txt

# Core ML
numpy==2.3.4
pandas==2.3.3
scikit-learn==1.8.0
xgboost==3.2.0
scipy==1.17.0
joblib==1.5.3

# Explainability
shap==0.51.0

# Dimensionality reduction + clustering
umap-learn==0.5.12
python-louvain==0.16

# Knowledge graph
networkx==3.6.1

# Dashboard & visualization
streamlit==1.56.0
plotly==6.7.0
matplotlib==3.10.8

# Data parsing
openpyxl==3.1.5
lxml==6.0.2
python-docx==1.2.0
tqdm==4.67.1

# Optional (obonet for HPO OBO parsing)
# obonet>=0.3.1
